consolidate rust modules, add basic docs skeleton
