+++ docs/api/index.md
@@ -1,4 +1,4 @@
-# API
+# noob
 
 ## Main modules
