+++ docs/api/noob-core/index.md
@@ -1,0 +1,14 @@
+# noob-core
+
+```{toctree}
+:maxdepth: 1
+
+lib
+bridge
+epoch
+exceptions
+item
+scheduler
+sorter
+tube
+```
