+++ docs/api/noob-core/lib.md
@@ -1,0 +1,49 @@
+# Crate `noob_core`
+
+:::::{rust:crate} noob_core
+:index: 0
+
+  :::
+  :::
+
+:::{rubric} Modules
+:::
+:::{toctree}
+:maxdepth: 1
+
+bridge
+epoch
+exceptions
+item
+scheduler
+sorter
+tube
+:::
+:::{rust:use} noob_core
+:used_name: self
+
+:::
+:::{rust:use} noob_core
+:used_name: crate
+
+:::
+
+:::{rubric} Types
+:::
+
+::::{rust:type} noob_core::FxIndexMap
+:index: 0
+:vis: pub
+:layout: [{"type":"keyword","value":"type"},{"type":"space"},{"type":"name","value":"FxIndexMap"},{"type":"punctuation","value":"<"},{"type":"name","value":"K"},{"type":"punctuation","value":", "},{"type":"name","value":"V"},{"type":"punctuation","value":">"}]
+
+  :::
+  :::
+::::
+::::{rust:type} noob_core::FxIndexSet
+:index: 0
+:vis: pub
+:layout: [{"type":"keyword","value":"type"},{"type":"space"},{"type":"name","value":"FxIndexSet"},{"type":"punctuation","value":"<"},{"type":"name","value":"T"},{"type":"punctuation","value":">"}]
+
+  :::
+  :::
+::::
