+++ docs/api/noob-core/tube.md
@@ -1,0 +1,27 @@
+# `mod tube`
+
+::::{rust:module} noob_core::tube
+:index: 0
+:vis: pub
+
+  :::
+  Tube helpers from rust :)
+  Just mirroring the python structure for now for ported code
+  :::
+:::{rust:use} noob_core::tube
+:used_name: self
+
+:::
+:::{rust:use} noob_core
+:used_name: crate
+
+:::
+:::{rust:use} noob_core::FxIndexSet
+:used_name: FxIndexSet
+
+:::
+:::{rust:use} noob_core::sorter::EdgeRec
+:used_name: EdgeRec
+
+:::
+::::
