+++ pyproject.toml
@@ -2,11 +2,15 @@
 dependencies = []
 requires-python = ">=3.11"
 
-[tool.pdm.dev-dependencies]
+[dependency-groups]
+
 dev = [
     "copier>=9.12.0",
-    "-e file:///${PROJECT_ROOT}/packages/noob[dev]",
-    "-e file:///${PROJECT_ROOT}/packages/nobes[dev]",
+    "noob[dev] @ file:///${PROJECT_ROOT}/packages/noob",
+    "nobes[dev] @ file:///${PROJECT_ROOT}/packages/nobes",
+]
+docs = [
+    "sphinxcontrib-rust>=1.2.0",
 ]
 
 [tool.pdm.resolution.overrides]
